requirements.txt
fastapi
uvicorn
cryptography
argon2-cffi
pydantic
python-multipart
slowapi
